third_party/WebKit/LayoutTests/inspector/sources/debugger/multi-target-breakpoint-manager-expected.txt

Tests BreakpointManager class.


Running: testResolvedBreakpointInBothTarget
  Created breakpoints manager
  Dumping Storage
  Adding script: a.js
  Adding UISourceCode: a.js
  Setting breakpoint at a.js:30:0 enabled:true condition:
    breakpointAdded(a.js, 30, 0, , true)
    target 1 debuggerModel.setBreakpoint(a.js:30:)
    breakpointRemoved(a.js, 30, 0)
    breakpointAdded(a.js, 30, 0, , true)
    Location created: target 1 a.js:30
    breakpointAdded(a.js, 131, 0, , true)
    Location created: target 1 a.js:131
  Dumping Breakpoint Locations
    UISourceCode (url='a.js', uri='a.js')
      Location: (30, 0)
      Location: (131, 0)
  Second target added
    target 2 debuggerModel.setBreakpoint(a.js:30:)
    Location created: target 2 a.js:30
  Dumping Breakpoint Locations
    UISourceCode (url='a.js', uri='a.js')
      Location: (30, 0)
      Location: (131, 0)
  First target removed
    breakpointRemoved(a.js, 131, 0)
    Location disposed: target 1 a.js:30
    Location disposed: target 1 a.js:131
  Dumping Breakpoint Locations
    UISourceCode (url='a.js', uri='a.js')
      Location: (30, 0)
  Second target removed
    breakpointRemoved(a.js, 30, 0)
    breakpointAdded(a.js, 30, 0, , true)
    Location disposed: target 2 a.js:30
  Dumping Breakpoint Locations
    UISourceCode (url='a.js', uri='a.js')
      Location: (30, 0)
  Dumping Storage
    a.js:30 enabled:true condition:
  Resetting breakpoint manager
    breakpointRemoved(a.js, 30, 0)
  Dumping Breakpoint Locations

Running: testResolvedBreakpointInOneTarget
  Created breakpoints manager
  Dumping Storage
  Adding script: a.js
  Adding UISourceCode: a.js
  Setting breakpoint at a.js:30:0 enabled:true condition:
    breakpointAdded(a.js, 30, 0, , true)
    target 1 debuggerModel.setBreakpoint(a.js:30:)
    target 2 debuggerModel.setBreakpoint(a.js:30:)
    breakpointRemoved(a.js, 30, 0)
    breakpointAdded(a.js, 30, 0, , true)
    Location created: target 1 a.js:30
  First target removed!
    breakpointRemoved(a.js, 30, 0)
    breakpointAdded(a.js, 30, 0, , true)
    Location disposed: target 1 a.js:30
  Dumping Breakpoint Locations
    UISourceCode (url='a.js', uri='a.js')
      Location: (30, 0)
  Dumping Breakpoint Locations
    UISourceCode (url='a.js', uri='a.js')
      Location: (30, 0)
  Dumping Storage
    a.js:30 enabled:true condition:
  Resetting breakpoint manager
    breakpointRemoved(a.js, 30, 0)
    target 2 debuggerModel.removeBreakpoint(a.js:30)
  Dumping Breakpoint Locations
